+++ docs/changelog.md
@@ -31,6 +31,7 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 - The command line flag `--reload-icon-themes` which can be used to tell a running Kando instance to reload all icon themes.
 - A setting to **enable or disable the drawing of selection wedges** in supported menu themes. This can be found in the theme-settings dialog. It is disabled by default as the selection wedges are quite resource-intensive to draw currently.
 - Many translation updates: **Thanks to all the contributors!**
+- Add support for GNOME's classic mode.
 
 ### :wrench: Changed